scripts/requirements.txt
# Migration Script Requirements
pandas>=1.5.0
mysql-connector-python>=8.0.0
sqlalchemy>=1.4.0